Constructing Statement Logic

+++ ASTNode/ASTExpressionNode/Binary/ExprBinOpAdd.h
@@ -16,7 +16,7 @@ namespace AST {
 
         ~ExprBinOpAdd() override = default;
 
-        double solve() const override {
+        float solve() const override {
             return get_pArg1()->solve() + get_pArg2()->solve();
         }
