+++ ASTNode/ASTExpressionNode/Data/ExprVar.h
@@ -5,7 +5,7 @@
 #ifndef COMPILER_ASTEXPRVAR_H
 #define COMPILER_ASTEXPRVAR_H
 
-#include "Expr.h"
+#include "../Expr.h"
 #include "Var.h"
 
 namespace AST {
@@ -16,7 +16,7 @@ namespace AST {
         Var *get_pVar() const {return _pVar;}
         explicit ExprVar(Var *pVar): Expr(), _pVar(pVar) { }
         ~ExprVar() override = default;
-        double solve() const override { return _pVar->get(); }
+        float solve() const override { return _pVar->get(); }
         void Accept(Visitor& v) override {
             v.Visit(*this);
         };
