+++ ASTNode/ASTExpressionNode/Expr.h
@@ -14,7 +14,7 @@ namespace AST {
         Expr() = default;
     public:
         virtual ~Expr() = default;
-        virtual double solve() const = 0;
+        virtual float solve() const = 0;
 
         virtual void Accept(Visitor& v) = 0;
     };
